Refactored DB Tests
Got rid of the giant file. NOTE: As of this commit, the old DB test
file, tst_drc_db_tests.cpp, is still there. This file is now deprecated
and should be removed as soon as possible. However, it is still there as
of this commit in order to pull changes, but will be removed/commented
out in a future commit