Added Lorenzo's tests to the refactored DB tests
Lorenzo's user table tests are now a part of the refactored tests. The
unit tests are much easier to read and set up. As of now, the old test
file, tst_drc_db_tests.cpp, is outdated and no longer necessary. I have
commented it out, and it is safe to remove that file